\name{adhan-package}
\alias{adhan-package}
\alias{adhan}
\docType{package}
\title{
adhan
}
\description{
The package provides functions to get Adhan (islamic prayer) times. It also provides functions to map prayer times of a city to another city.
}
\details{

License: GPL-2
}
\author{
Ali Oghabian <Ali.Oghabian@Helsinki.Fi>

Maintainer: Ali Oghabian <Ali.Oghabian@Helsinki.Fi>
}
\references{
https://aladhan.com
}
\keyword{ package }
